requirements.txt
# Core dependencies
psycopg[binary]==3.2.11
sqlalchemy==2.0.44
alembic==1.17.0
setuptools<81.0.0

# Audio processing
librosa==0.11.0
scipy==1.16.2
numpy==2.3.4
soundfile==0.13.1

# Video processing
yt-dlp==2025.10.14
ffmpeg-python==0.2.0

# Social media APIs
tweepy==4.16.0
praw==7.8.1

# Google APIs
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Web server for OAuth callbacks
fastapi==0.119.0
uvicorn==0.38.0
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.2.0

# API Authentication & Security
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# API Rate Limiting
slowapi==0.1.9

# Utilities
python-dotenv==1.1.1
requests==2.32.5
python-dateutil==2.9.0.post0
tqdm==4.67.1
colorlog==6.10.1

# Observability
prometheus-client==0.21.1

# Development
pytest==8.4.2
black==25.9.0
flake8==7.3.0
httpx==0.28.1  # For FastAPI testing

# Async support
aiohttp==3.13.1
